Basic post browser functionality; thumbnail loading, selection, info bar.

+++ Booru-chan/Booru-chan/InfoBarController.swift
@@ -22,7 +22,7 @@ class InfoBarController: NSViewController {
         }
     }
 
-    var loadingProgress: Float! {
+    var loadingProgress: Double! {
         didSet {
             updateLabel();
         }
